requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
black==26.3.1
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
databricks-sdk==0.102.0
distlib==0.4.0
docker==7.1.0
fastapi==0.135.3
filelock==3.25.2
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.2
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huey==2.6.0
identify==2.6.18
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.5.3
kiwisolver==1.5.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mlflow==3.11.1
mlflow-skinny==3.11.1
mlflow-tracing==3.11.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
nodeenv==1.10.0
numpy==2.4.3
opencv-python==4.13.0.92
opentelemetry-api==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
packaging==26.0
pandas==2.3.3
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.6
pluggy==1.6.0
polars==1.39.3
polars-runtime-32==1.39.3
pre_commit==4.5.1
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
pytest==9.0.3
python-dateutil==2.9.0.post0
python-discovery==1.2.2
python-dotenv==1.2.2
python-multipart==0.0.26
pytokens==0.4.1
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
skops==0.13.0
smmap==5.0.3
SQLAlchemy==2.0.49
sqlparse==0.5.5
starlette==1.0.0
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
ultralytics==8.4.37
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.44.0
virtualenv==21.2.1
waitress==3.0.2
wcwidth==0.6.0
Werkzeug==3.1.8
yarl==1.23.0
zipp==3.23.0
